[product] feature: authorization with JWT claim (#24)
- extract quota / permission data from JWT claims , to validate user authorization
- permission check and saleable item quota check on web API endpoints
- add new repository interface for retrieving number of saleable items
  saved for specific seller

+++ services/product/v2/src/product/api/web/attribute_label.py
@@ -3,12 +3,19 @@ from typing import List, Optional
 from blacksheep import Response, Request, Content
 from blacksheep.server.authorization import auth
 from blacksheep.server.controllers import APIController
-from blacksheep.server.responses import created, ok, no_content  # status_code
+from blacksheep.server.responses import (
+    created,
+    ok,
+    forbidden,
+    no_content,
+)  # status_code
 from blacksheep.server.bindings import FromJSON, FromQuery, QueryBinder, BoundValue
+from guardpost import User as AuthUser
 import pydantic_core
 
 from product.model import AttrLabelModel
 from product.shared import SharedContext
+from product.util import PriviledgeLevel, permission_check
 
 from . import router
 from ..dto import AttrCreateReqDto, AttrUpdateReqDto
@@ -27,11 +34,17 @@ class ListStrQueryBinder(QueryBinder):
 
 
 class AttrLabelController(APIController):
-    @auth("authed_staff_only")
+    @auth(PriviledgeLevel.AuthedUser.value)
     @router.post("/attributes")
     async def create(
-        self, shr_ctx: SharedContext, reqbody: FromJSON[List[AttrCreateReqDto]]
+        self,
+        shr_ctx: SharedContext,
+        reqbody: FromJSON[List[AttrCreateReqDto]],
+        user: AuthUser,
     ) -> Response:
+        perm_err = permission_check(user.claims, ["add_productattributetype"])
+        if perm_err:
+            return forbidden(message=perm_err)
         reqbody = reqbody.value
         repo = shr_ctx.datastore.prod_attri
         ms = AttrLabelModel.from_create_reqs(reqbody)
@@ -39,11 +52,17 @@ class AttrLabelController(APIController):
         respbody = AttrLabelModel.to_resps(ms)
         return created(message=respbody)
 
-    @auth("authed_staff_only")
+    @auth(PriviledgeLevel.AuthedUser.value)
     @router.put("/attributes")
     async def update(
-        self, shr_ctx: SharedContext, reqbody: FromJSON[List[AttrUpdateReqDto]]
+        self,
+        shr_ctx: SharedContext,
+        reqbody: FromJSON[List[AttrUpdateReqDto]],
+        user: AuthUser,
     ) -> Response:
+        perm_err = permission_check(user.claims, ["change_productattributetype"])
+        if perm_err:
+            return forbidden(message=perm_err)
         reqbody = reqbody.value
         repo = shr_ctx.datastore.prod_attri
         ms = AttrLabelModel.from_update_reqs(reqbody)
@@ -51,9 +70,14 @@ class AttrLabelController(APIController):
         respbody = AttrLabelModel.to_resps(ms)
         return ok(message=respbody)
 
-    @auth("authed_staff_only")
+    @auth(PriviledgeLevel.AuthedUser.value)
     @router.delete("/attributes")
-    async def delete(self, shr_ctx: SharedContext, ids: FromQueryListStr) -> Response:
+    async def delete(
+        self, shr_ctx: SharedContext, ids: FromQueryListStr, user: AuthUser
+    ) -> Response:
+        perm_err = permission_check(user.claims, ["delete_productattributetype"])
+        if perm_err:
+            return forbidden(message=perm_err)
         repo = shr_ctx.datastore.prod_attri
         await repo.delete(ids=ids.value)
         return no_content()
